+++ darts/models/forecasting/ensemble_model.py
@@ -3,9 +3,15 @@ Ensemble Model Base Class
 """
 
 import os
+import sys
 from abc import abstractmethod
 from collections.abc import Sequence
 from typing import BinaryIO, Optional, Union
+
+if sys.version_info >= (3, 11):
+    from typing import Self
+else:
+    from typing_extensions import Self
 
 from darts.logging import get_logger, raise_if, raise_if_not, raise_log
 from darts.models.forecasting.forecasting_model import (
@@ -386,7 +392,7 @@ class EnsembleModel(GlobalForecastingModel):
             ]
         return predictions[0] if is_single_series else predictions
 
-    def _clean(self) -> "EnsembleModel":
+    def _clean(self) -> Self:
         """Cleans the model and sub-models."""
         cleaned_model = super()._clean()
         cleaned_model.forecasting_models = [
@@ -433,8 +439,11 @@ class EnsembleModel(GlobalForecastingModel):
             checkpoint) are saved under ``"{path}.{ithModelClass}_{i}.pt"`` and ``"{path}.{ithModelClass}_{i}.ckpt"``.
         clean
             Whether to store a cleaned version of the model. If `True`, the training series and covariates are removed.
-            Note: After loading the model, a `series` must be passed 'predict()', `historical_forecasts()` and other
-            forecasting methods.
+            If the underlying `forecasting_models` contain any `TorchForecastingModel`, will additionally remove all of
+            their Lightning Trainer-related parameters.
+
+            Note: After loading a model stored with `clean=True`, a `series` must be passed 'predict()',
+            `historical_forecasts()` and other forecasting methods.
         pkl_kwargs
             Keyword arguments passed to `pickle.dump()`
         """
@@ -457,23 +466,24 @@ class EnsembleModel(GlobalForecastingModel):
         **kwargs,
     ) -> "EnsembleModel":
         """
-        Loads an ensemble model from a given path.
+        Loads a model from a given path or file handle.
 
         Parameters
         ----------
         path
-            Path from which to load the model.
+            Path or file handle from which to load the model.
         pl_trainer_kwargs
-            Optionally, a set of kwargs to create the new PyTorch Lightning Trainers
-            used to handle the ``TorchForecastingModel``.
-            Running on GPU(s) is also possible using ``pl_trainer_kwargs`` by specifying keys ``"accelerator",
-            "devices", and "auto_select_gpus"``.
-            Check the `Lightning Trainer documentation <https://pytorch-lightning.readthedocs.io/en/stable/common/trainer.html>`_
+            Only effective if the underlying forecasting models contain a `TorchForecastingModel`.
+            Optionally, a set of kwargs to create a new Lightning Trainer used to configure the model for downstream
+            tasks (e.g. prediction).
+            Some examples include specifying the batch size or moving the model to CPU/GPU(s). Check the
+            `Lightning Trainer documentation <https://pytorch-lightning.readthedocs.io/en/stable/common/trainer.html>`_
             for more information about the supported kwargs.
         **kwargs
+            Only effective if the underlying forecasting models contain a `TorchForecastingModel`.
             Additional kwargs for PyTorch Lightning's :func:`LightningModule.load_from_checkpoint()` method,
             For more information, read the `official documentation <https://pytorch-lightning.readthedocs.io/en/stable/
-            common/lightning_module.html#load-from-checkpoint>`_. This is only used for ``TorchForecastingModel``.
+            common/lightning_module.html#load-from-checkpoint>`_.
         """
         model: EnsembleModel = GlobalForecastingModel.load(path)