+++ darts/models/forecasting/torch_forecasting_model.py
@@ -28,6 +28,11 @@ from collections.abc import Sequence
 from glob import glob
 from typing import Any, Callable, Literal, Optional, Union
 
+if sys.version_info >= (3, 11):
+    from typing import Self
+else:
+    from typing_extensions import Self
+
 import numpy as np
 import pandas as pd
 import pytorch_lightning as pl
@@ -1643,7 +1648,7 @@ class TorchForecastingModel(GlobalForecastingModel, ABC):
                 aggregated.append([sample[i] for sample in batch])
         return tuple(aggregated)
 
-    def _clean(self):
+    def _clean(self) -> Self:
         """Returns a cleaned model, keeping only the necessary attributes for prediction."""
         model = super()._clean()
         # Copy from super()._clean() call __getstate__ which removes model and trainer
@@ -1688,10 +1693,12 @@ class TorchForecastingModel(GlobalForecastingModel, ABC):
             is automatically saved under ``"{ModelClass}_{YYYY-mm-dd_HH_MM_SS}.pt"``.
             E.g., ``"RNNModel_2020-01-01_12_00_00.pt"``.
         clean
-            Whether to store a cleaned version of the model. If `True`, only model parameters
-            and attributes needed for prediction are saved.
-            Note: After loading the model, a `series` must be passed 'predict()', `historical_forecasts()` and other
-            forecasting methods.
+            Whether to store a cleaned version of the model. If `True`, the training series and covariates are removed.
+            Additionally, removes all Lightning Trainer-related parameters (passed with `pl_trainer_kwargs` at model
+            creation).
+
+            Note: After loading a model stored with `clean=True`, a `series` must be passed 'predict()',
+            `historical_forecasts()` and other forecasting methods.
         """
         if path is None:
             # default path
@@ -1703,7 +1710,7 @@ class TorchForecastingModel(GlobalForecastingModel, ABC):
         with open(path, "wb") as f_out:
             torch.save(model_to_save, f_out)
 
-        # save the LightningModule checkpoint (only model weights on clean)
+        # save the LightningModule checkpoint (weights only with `clean=True`)
         path_ptl_ckpt = path + ".ckpt"
         if self.trainer is not None:
             self.trainer.save_checkpoint(path_ptl_ckpt, weights_only=clean)
@@ -1738,16 +1745,26 @@ class TorchForecastingModel(GlobalForecastingModel, ABC):
                 model_loaded = RNNModel.load(path)
             ..
 
+        Example for loading an :class:`RNNModel` to GPU:
+
+            .. highlight:: python
+            .. code-block:: python
+
+                from darts.models import RNNModel
+
+                model_loaded = RNNModel.load(path, pl_trainer_kwargs={"accelerator": "gpu"})
+            ..
+
         Parameters
         ----------
         path
             Path from which to load the model. If no path was specified when saving the model, the automatically
             generated path ending with ".pt" has to be provided.
         pl_trainer_kwargs
-            Optionally, a set of kwargs to create a new PyTorch Lightning Trainer used to handle the underlying module.
-            Running on GPU(s) is also possible using ``pl_trainer_kwargs`` by specifying keys ``"accelerator",
-            "devices", and "auto_select_gpus"``.
-            Check the `Lightning Trainer documentation <https://pytorch-lightning.readthedocs.io/en/stable/common/trainer.html>`_
+            Optionally, a set of kwargs to create a new Lightning Trainer used to configure the model for downstream
+            tasks (e.g. prediction).
+            Some examples include specifying the batch size or moving the model to CPU/GPU(s). Check the
+            `Lightning Trainer documentation <https://pytorch-lightning.readthedocs.io/en/stable/common/trainer.html>`_
             for more information about the supported kwargs.
         **kwargs
             Additional kwargs for PyTorch Lightning's :func:`LightningModule.load_from_checkpoint()` method,
@@ -1769,7 +1786,7 @@ class TorchForecastingModel(GlobalForecastingModel, ABC):
                 f"found at {path_ptl_ckpt}. Please call `fit()` before calling `predict()`."
             )
 
-        # if the model was saved with clean=True or if pl_trainer_kwargs is provided
+        # if the model was saved with `clean=True` or if pl_trainer_kwargs is provided
         # we overwrite the trainer params
         model.trainer_params = (
             {"callbacks": [], "accelerator": "cpu"}  # by default, use CPU